+++ database/migrations/2016_05_21_012933_create_ingredients_table.php
@@ -12,7 +12,7 @@ class CreateIngredientsTable extends Migration {
     public function up() {
         Schema::create('ingredients', function (Blueprint $table) {
             $table->increments('id');
-            $table->bigInteger('id_recipe')->unsigned();
+            $table->integer('id_recipe')->unsigned();
             $table->integer('quantity');
             $table->integer('measure')->default(1);
             $table->longText('description');
